+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - Minor: Implement static `Calendar` component
+  - Minor: Implement `Icon` v2 API
+  - Minor: Implement `IconButton` component
 </details>
 
 ## v0.35.0 (August 5, 2020)